Extract TenantManager out of pusher into its own package
The intention here is that it should be possible to use the tenant
manager in more places than just the pusher, and it should be possible
to share the cache.
